Move run_anvil_audit management command to auditor app
